# openapi_spec.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "AiparatiExpress API",
        "description": "API automatizada para candidaturas Portugal 2030",
        "version": "1.0.0"
    },
    "paths": {
        "/": {
            "get": {
                "summary": "Root",
                "description": "Health check",
                "operationId": "root__get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/upload": {
            "post": {
                "summary": "Upload Ies",
                "description": "Upload de ficheiro IES para processamento\n\n- **file**: Ficheiro PDF da IES\n- **context**: Contexto adicional opcional\n\nRetorna task_id para acompanhamento",
                "operationId": "upload_ies_api_upload_post",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_upload_ies_api_upload_post"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProcessResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/api/status/{task_id}": {
            "get": {
                "summary": "Get Status",
                "description": "Verifica status de processamento",
                "operationId": "get_status_api_status__task_id__get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "task_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Task Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/result/{task_id}": {
            "get": {
                "summary": "Get Result",
                "description": "Obt\u00e9m resultado completo da an\u00e1lise",
                "operationId": "get_result_api_result__task_id__get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "task_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Task Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AnalysisResult"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/download/{task_id}/{file_type}": {
            "get": {
                "summary": "Download File",
                "description": "Download de ficheiros gerados",
                "operationId": "download_file_api_download__task_id___file_type__get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "task_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Task Id"
                        }
                    },
                    {
                        "name": "file_type",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "File Type"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/tasks": {
            "get": {
                "summary": "List Tasks",
                "description": "Lista tarefas do utilizador",
                "operationId": "list_tasks_api_tasks_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/api/tasks/{task_id}": {
            "delete": {
                "summary": "Delete Task",
                "description": "Elimina tarefa e ficheiros associados",
                "operationId": "delete_task_api_tasks__task_id__delete",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "task_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Task Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "summary": "Health Check",
                "description": "Health check detalhado",
                "operationId": "health_check_health_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "AnalysisResult": {
                "properties": {
                    "metadata": {
                        "type": "object",
                        "title": "Metadata"
                    },
                    "dados_financeiros": {
                        "type": "object",
                        "title": "Dados Financeiros"
                    },
                    "analise": {
                        "type": "object",
                        "title": "Analise"
                    },
                    "download_urls": {
                        "additionalProperties": {
                            "type": "string"
                        },
                        "type": "object",
                        "title": "Download Urls"
                    }
                },
                "type": "object",
                "required": [
                    "metadata",
                    "dados_financeiros",
                    "analise",
                    "download_urls"
                ],
                "title": "AnalysisResult"
            },
            "Body_upload_ies_api_upload_post": {
                "properties": {
                    "file": {
                        "type": "string",
                        "format": "binary",
                        "title": "File"
                    },
                    "nif": {
                        "type": "string",
                        "title": "Nif"
                    },
                    "ano_exercicio": {
                        "type": "string",
                        "title": "Ano Exercicio"
                    },
                    "designacao_social": {
                        "type": "string",
                        "title": "Designacao Social"
                    },
                    "email": {
                        "type": "string",
                        "title": "Email"
                    },
                    "context": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Context"
                    }
                },
                "type": "object",
                "required": [
                    "file",
                    "nif",
                    "ano_exercicio",
                    "designacao_social",
                    "email"
                ],
                "title": "Body_upload_ies_api_upload_post"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "ProcessResponse": {
                "properties": {
                    "task_id": {
                        "type": "string",
                        "title": "Task Id"
                    },
                    "status": {
                        "type": "string",
                        "title": "Status"
                    },
                    "message": {
                        "type": "string",
                        "title": "Message"
                    }
                },
                "type": "object",
                "required": [
                    "task_id",
                    "status",
                    "message"
                ],
                "title": "ProcessResponse"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}